REFS : #904 fixed slider min max by converting them to int

+++ grails-app/services/species/trait/TraitService.groovy
@@ -429,7 +429,7 @@ class TraitService extends AbstractObjectService {
 
         Sql sql = Sql.newInstance(dataSource);
         List numericTraitMinMax =  sql.rows("""
-        select min(f.value::float),max(f.to_value::float),t.id from fact f,trait t where f.trait_id = t.id and t.data_types='NUMERIC' group by t.id;
+        select min(f.value::float)::integer,max(f.to_value::float)::integer,t.id from fact f,trait t where f.trait_id = t.id and t.data_types='NUMERIC' group by t.id;
         """);
         return [instanceList:instanceList, instanceTotal:allInstanceCount, queryParams:queryParts.queryParams, activeFilters:queryParts.activeFilters, 'traitFactMap':queryParts.traitFactMap, 'object':queryParts.object,numericTraitMinMax:numericTraitMinMax];
     }
